bringing up to date

+++ gr-analog/include/gnuradio/analog/frequency_modulator_fc.h
@@ -28,13 +28,36 @@
 
 namespace gr {
   namespace analog {
-    
+
     /*!
      * \brief Frequency modulator block
      * \ingroup modulators_blk
      *
      * \details
      * float input; complex baseband output
+     *
+     * Takes a real, baseband signal (x_m[n]) and output a frequency
+     * modulated signal (y[n]) according to:
+     *
+     * \f[
+     * y[n] = exp (j 2 \pi \frac{f_{\Delta}}{f_s} \sum{x[n]})
+     * \f]
+     *
+     * Where x[n] is the input sample at time n and \f$ f_{\Delta} \f$
+     * is the frequency deviation. Common values for \f$ f_{\Delta}
+     * \f$ are 5 kHz for narrowband FM channels such as for voice
+     * systems and 75 KHz for wideband FM, like audio broadcast FM
+     * stations.
+     *
+     * In this block, the input argument is \p sensitivity, not the
+     * frequency deviation.  The sensitivity specifies how much the
+     * phase changes based on the new input sample. Given a maximum
+     * deviation, \f$ f_{\Delta} \f$, and sample rate \f$f_s\f$, the
+     * sensitivity is defined as:
+     *
+     * \f[
+     * k = 2 \pi \frac{f_{\Delta}}{f_s}
+     * \f]
      */
     class ANALOG_API frequency_modulator_fc : virtual public sync_block
     {
@@ -47,7 +70,7 @@ namespace gr {
        *
        * \param sensitivity radians/sample = amplitude * sensitivity
        */
-      static sptr make(double sensitivity);
+      static sptr make(float sensitivity);
 
       virtual void set_sensitivity(float sens) = 0;
       virtual float sensitivity() const = 0;
